Custom authorizer now works
This commit implements a working custom authorizer which only yields
access when the appropriate secrets are provided as expected in
the request.

To enable this functionality, it also introduces several new roles
which permit the Lambda functions involved to access the secret in
AWS SecretsManager which contains the auth token generated for
Okta to provide to this application.