Authorizer: Cleanup + (limited) caching
This commit refactors the authorizer code to be a little bit cleaner.
It also implements limited caching of the secret which relies on the
fact that AWS Lambda containers are reused when possible.

At this time, caching will only function for reused application
containers. If the cache is more than 300 seconds old, it will
be refreshed.