Implement verification + remove some boiler plate
Event hooks in Okta require a verified endpoint. Verification makes
a GET request to the endpoint provided and expects to receive back
the verification string provided in the X-Okta-Verification-Challenge
header.

Primarily this commit implements the ability to verify an endpoint with
Okta (infrastructure and code). It also removes some boiler plate code
added with the "hello world" project.