Add SECRET_ID env variable to avoid hardcoding
While there was previously no secret hardcoded in the code, there was a
hardcoded name reference which could be broken if the secret name were
to be changed.

This updates ensures that the secret name is dynamically provided to AWS
Lambda so that it will always look up the desired secret.